requirements.txt
flask==3.0.0
requests==2.31.0
httpx<0.27.0
openai==1.3.0
pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
google-generativeai==0.8.5
newsapi-python>=0.2.6
scikit-learn>=1.0.0
psycopg2-binary
cloudinary